requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
pdf2image
pypdf2
python-docx
openpyxl
jinja2
python-magic
qrcode